Support decimal numbers in lexical and syntactic analysis
Updated the lexer to correctly parse and handle decimal numbers by modifying `TokenType`, `Token`, and `AnaliseLexica` classes. Adjusted the parser and abstract syntax tree to process `Double` instead of `int` for numeric values. Added necessary safeguards for decimal input formatting.

+++ ArvoreSintatica.java
@@ -3,8 +3,9 @@ class ArvoreSintatica{}
 class Exp extends ArvoreSintatica{}
 
 class Num extends Exp{
- 	int num;
-	Num(int num){this.num=num;}
+ 	Double num;
+	Num(Double num){this.num=num;}
+
 }
 
 class Operador extends Exp{
